refractor: move components out of react subfolder
