+++ src/components/ModelSelector.js
@@ -8,9 +8,9 @@ import {
   ToggleButton,
 } from "react-bootstrap";
 
-import { useSelectorContext } from "../../context/selector-context";
-import { myChangeColorMapAction as changeColorMap } from "../../redux/AppActions";
-import { colorMaps } from "../../assets/config.json";
+import { useSelectorContext } from "../context/selector-context";
+import { myChangeColorMapAction as changeColorMap } from "../redux/AppActions";
+import { colorMaps } from "../assets/config.json";
 
 export default function ModelSelector(props) {
   const { sidebarVisible, setSidebarVisible } = props;
