+++ src/components/controls/ColorMapControl.js
@@ -6,9 +6,9 @@ import "primereact/resources/primereact.min.css";
 import "primeicons/primeicons.css";
 
 import { connect } from "react-redux";
-import { myChangeColorMapAction as changeColorMap } from "../../../redux/AppActions";
+import { myChangeColorMapAction as changeColorMap } from "../../redux/AppActions";
 
-import { colorMaps } from "../../../assets/config.json";
+import { colorMaps } from "../../assets/config.json";
 
 const mapStateToProps = (state) => {
   return { colorMap: state.colorMap };
